+++ svc/api/cloud/src/route/games/tokens.rs
@@ -10,7 +10,9 @@ pub async fn create_cloud_token(
 	game_id: Uuid,
 	_body: serde_json::Value,
 ) -> GlobalResult<models::CloudGamesCreateCloudTokenResponse> {
-	ctx.auth().check_game_write(ctx.op_ctx(), game_id).await?;
+	ctx.auth()
+		.check_game_write_or_admin(ctx.op_ctx(), game_id)
+		.await?;
 
 	let token_res = op!([ctx] cloud_game_token_create {
 		game_id: Some(game_id.into()),